python/pyproject.toml
[project]
name = "binance-expert"
version = "0.1.0"
description = "Application for Binance expert"
authors = [{ name = "Oleg Alexandrov", email = "alexandrovoleg.ru@gmail.com" }]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "binance-connector>=3.12.0",
    "flask>=3.1.0",
    "numpy>=2.2.2",
    "pandas>=2.2.3",
]

[dependency-groups]
dev = [
    "mypy>=1.15.0",
    "python-dotenv-vault>=0.6.4",
    "ruff>=0.9.5",
]
test = [
    "pytest>=8.3.4",
]
docs = [
    "pytest>=8.3.4",
]

[tool.uv]
package = false
link-mode = "copy"
default-groups = [
    "test",
    "docs",
]
